% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mmcm.R
\name{print.mmcm}
\alias{print.mmcm}
\title{Print function for mmcm object}
\usage{
\method{print}{mmcm}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{Object of class \code{mmcm}, which is result of function \code{\link{mcm.mvt}}, \code{\link{mmcm.mvt}} and \code{\link{mmcm.resamp}}.}

\item{digits}{a non-null value for digits specifies the minimum number of significant digits to be printed in values. The default, NULL, uses \code{\link[base:options]{getOption}}(digits). (For the interpretation for complex numbers see \code{\link[base:Round]{signif}}.) Non-integer values will be rounded down, and only values greater than or equal to 1 and no greater than 22 are accepted.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
This function print result of function \code{\link{mcm.mvt}}, \code{\link{mmcm.mvt}} and \code{\link{mmcm.resamp}}
}
\details{
The case where printed "More than 2 contrast coefficient vectors were selected", some contrast may be unsuitable.
}
\seealso{
\code{\link[base:print]{print.default}},
\code{\link{mmcm.mvt}},
\code{\link{mmcm.resamp}},
\code{\link{mcm.mvt}}
}
\keyword{print}

